model/requirements.txt
torch==2.3.1
transformers==4.42.4
pillow==10.3.0
# flash_attn==2.5.8
accelerate==0.30.1
loguru~=0.7.2
sentencepiece==0.2.0
timm==1.0.3
google-cloud-storage==2.17.0
celery[redis]==5.4.0
einops==0.8.0